Refactored the LoggerTree and removed the Crashlytics dependency from the core
