+++ app/src/main/java/com/xmartlabs/template/ApplicationComponent.java
@@ -4,7 +4,6 @@ import com.xmartlabs.base.core.controller.Controller;
 import com.xmartlabs.base.core.controller.SharedPreferencesController;
 import com.xmartlabs.base.core.module.CoreAndroidModule;
 import com.xmartlabs.base.core.module.GsonModule;
-import com.xmartlabs.base.core.module.LoggerModule;
 import com.xmartlabs.base.core.module.OkHttpModule;
 import com.xmartlabs.base.core.module.PicassoModule;
 import com.xmartlabs.base.retrofit.module.RestServiceModule;
@@ -26,10 +25,8 @@ import dagger.Component;
 @Component(modules = {
     CoreAndroidModule.class,
     ControllerModule.class,
-    LoggerModule.class,
     GsonModule.class,
     ServiceGsonModule.class,
-    LoggerModule.class,
     OkHttpModule.class,
     PicassoModule.class,
     RestServiceModule.class,
